Add Markdown relative URL rewriting
When rendering MarkDown, use the pymdown-extensions pathconverter
extensions to rewrite relative links/images.

The rewritten URL is always of 'raw' view. In other words, when
we are rendering

  repo-name/blob/master/test.md

and that MarkDown document contains a relative link to an image,

  ![Test image](images/test.png)

then in the rendered HTML, we want the image source to be:

  <img src="repo-name/raw/master/images/test.png">

(note "/raw/" vs "/blob").

This is required for the image to display correctly -- the browser
will not be able to display repo-name/blob/master/images/test.png
which would be rendered to another HTML document, not an image.

In addition, this patch adds a rudimentary MIME type mapping
to raw view, which appears to be required for browsers (Safari
and Chrome, anyway) to render SVG images correctly.